Fix TypeError in device_tracker due to non-integer history_minutes
The `deque` `maxlen` argument requires an integer, but `history_minutes` could be a float (e.g. from config), causing a TypeError.

This change:
1. Casts `vehicle_position_history_minutes` to `int` in `__init__.py` when reading from config.
2. Explicitly casts the `maxlen` calculation to `int` in `device_tracker.py` as a defensive measure.
3. Adds a regression test.

+++ custom_components/gtfs_realtime/device_tracker.py
@@ -109,7 +109,7 @@ class GtfsVehicleTracker(TrackerEntity, CoordinatorEntity):
         # Position history for path drawing (configurable time window)
         history_minutes = coordinator.vehicle_position_history_minutes
         self._position_history: deque[tuple[float, float, datetime]] = deque(
-            maxlen=history_minutes * 2  # Assuming ~30s updates = 2 per minute
+            maxlen=int(history_minutes * 2)  # Assuming ~30s updates = 2 per minute
         )
 
     @property
